LPS-146932 Simplify BuildOutlet and create new child components
