+++ modules/apps/site-initializer/site-initializer-testray/extra/remote-app/src/pages/Project/Routines/Builds/BuildOutlet.tsx
@@ -13,8 +13,7 @@
  */
 
 import {useQuery} from '@apollo/client';
-import ClayChart from '@clayui/charts';
-import {useEffect, useRef} from 'react';
+import {useEffect} from 'react';
 import {
 	Outlet,
 	useLocation,
@@ -22,168 +21,19 @@ import {
 	useParams,
 } from 'react-router-dom';
 
-import Container from '../../../../components/Layout/Container';
-import QATable from '../../../../components/Table/QATable';
-import useTotalTestCases from '../../../../data/useTotalTestCases';
 import {
-	CType,
 	TestrayBuild,
+	TypePagination,
 	getTestrayBuild,
 } from '../../../../graphql/queries';
+import {
+	TestrayTask,
+	getTestrayTasks,
+} from '../../../../graphql/queries/testrayTask';
 import useHeader from '../../../../hooks/useHeader';
 import i18n from '../../../../i18n';
-import {getDonutLegend} from '../../../../util/graph';
-import {TotalTestCases} from '../../../../util/mock';
-
-type BuildOverviewProps = {
-	testrayBuild: TestrayBuild;
-};
-
-const BuildOverview: React.FC<BuildOverviewProps> = ({testrayBuild}) => {
-	const ref = useRef<any>();
-
-	const totalTestCases = useTotalTestCases();
-
-	return (
-		<>
-			<Container title={i18n.translate('details')}>
-				<QATable
-					items={[
-						{
-							title: i18n.translate('product-version'),
-							value: '7.0.x',
-						},
-						{
-							title: i18n.translate('description'),
-							value: testrayBuild.description,
-						},
-						{
-							title: i18n.translate('git-hash'),
-							value:
-								testrayBuild.gitHash ||
-								'c33e85e8b067d805a45956c76ad053ca98ffcc8a',
-						},
-						{
-							title: i18n.translate('create-date'),
-							value: testrayBuild.dateCreated,
-						},
-						{
-							title: i18n.translate('created-by'),
-							value: 'John Doe',
-						},
-						{title: i18n.translate('all-issues-found'), value: '-'},
-					]}
-				/>
-
-				<div className="d-flex mt-4">
-					<dl>
-						<dd>{i18n.sub('x-minutes', '0')}</dd>
-
-						<dd className="small-heading">
-							{i18n.translate('total-estimated-time')}
-						</dd>
-					</dl>
-
-					<dl className="ml-3">
-						<dd>{i18n.sub('x-minutes', '0')}</dd>
-
-						<dd className="small-heading">
-							{i18n.translate('total-estimated-time')}
-						</dd>
-					</dl>
-
-					<dl className="ml-3">
-						<dd>{i18n.sub('x-minutes', '0')}</dd>
-
-						<dd className="small-heading">
-							{i18n.sub('time-x-total-issues', '0')}
-						</dd>
-					</dl>
-				</div>
-			</Container>
-
-			<Container className="mt-4" title="Total Test Cases">
-				<div className="row">
-					<div className="col-2">
-						<ClayChart
-							data={{
-								colors: totalTestCases.colors,
-								columns: totalTestCases.donut.columns,
-								type: 'donut',
-							}}
-							donut={{
-								expand: false,
-								label: {
-									show: false,
-								},
-								legend: {
-									show: false,
-								},
-								title: totalTestCases.donut.total.toString(),
-								width: 15,
-							}}
-							legend={{show: false}}
-							onafterinit={() => {
-								getDonutLegend(ref.current, {
-									data: TotalTestCases.map(([name]) => name),
-									elementId: 'testrayTotalMetricsGraphLegend',
-									total: totalTestCases.donut.total as number,
-								});
-							}}
-							ref={ref}
-							size={{
-								height: 200,
-							}}
-						/>
-					</div>
-
-					<div className="col-2">
-						<div id="testrayTotalMetricsGraphLegend" />
-					</div>
-
-					<div className="col-8">
-						<ClayChart
-							axis={{
-								y: {
-									label: {
-										position: 'outer-middle',
-										text: i18n
-											.translate('tests')
-											.toUpperCase(),
-									},
-								},
-							}}
-							bar={{
-								width: {
-									max: 30,
-								},
-							}}
-							data={{
-								colors: totalTestCases.colors,
-								columns: totalTestCases.barChart.columns,
-								groups: [totalTestCases.statuses],
-								type: 'bar',
-							}}
-							legend={{
-								inset: {
-									anchor: 'top-right',
-									step: 1,
-									x: 10,
-									y: -20,
-								},
-								position: 'inset',
-							}}
-							padding={{
-								bottom: 5,
-								top: 20,
-							}}
-						/>
-					</div>
-				</div>
-			</Container>
-		</>
-	);
-};
+import BuildAlertBar from './BuildAlertBar';
+import BuildOverview from './BuildOverview';
 
 type BuildOutletProps = {
 	ignorePath: string;
@@ -193,18 +43,24 @@ const BuildOutlet: React.FC<BuildOutletProps> = ({ignorePath}) => {
 	const {pathname} = useLocation();
 	const {projectId, routineId, testrayBuildId} = useParams();
 	const {testrayProject, testrayRoutine}: any = useOutletContext();
-	const {data} = useQuery<CType<'testrayBuild', TestrayBuild>>(
-		getTestrayBuild,
-		{
-			variables: {
-				testrayBuildId,
-			},
-		}
+	const {data} = useQuery<{testrayBuild: TestrayBuild}>(getTestrayBuild, {
+		variables: {
+			testrayBuildId,
+		},
+	});
+
+	const {data: testrayTasksData} = useQuery<
+		TypePagination<'testrayTasks', TestrayTask>
+	>(getTestrayTasks);
+
+	const testrayBuild = data?.testrayBuild;
+	const testrayTasks = testrayTasksData?.testrayTasks.items || [];
+	const testrayTask = testrayTasks.find(
+		(testrayTask) =>
+			testrayTask?.testrayBuild?.id === Number(testrayBuildId)
 	);
 
 	const isCurrentPathIgnored = pathname.includes(ignorePath);
-
-	const testrayBuild = data?.c?.testrayBuild;
 
 	const basePath = `/project/${projectId}/routines/${routineId}/build/${testrayBuildId}`;
 
@@ -265,7 +121,13 @@ const BuildOutlet: React.FC<BuildOutletProps> = ({ignorePath}) => {
 		return (
 			<>
 				{!isCurrentPathIgnored && (
-					<BuildOverview testrayBuild={testrayBuild} />
+					<>
+						{testrayTask && (
+							<BuildAlertBar testrayTask={testrayTask} />
+						)}
+
+						<BuildOverview testrayBuild={testrayBuild} />
+					</>
 				)}
 
 				<Outlet />
